+++ alternativeMains/mainWebOld.go
@@ -1,4 +1,4 @@
-package main
+package alternativeMains
 
 import (
 	"html/template"
